2-strlen.c

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -5,7 +5,7 @@
 * @str: string.
 * Return: Length.
 */
-size_t _strlen(const char *str)
+int _strlen(const char *str)
 {
 size_t len = 0;
 while (*str++)
